Completed another iteration of the Kong palette loading routines
This is far less buggy than the previous method.  It hijacks the Kong palette loading routine used to set the inactive, 2P competition mode, invincibility, etc. palettes, and checks the kong_palette_order RAM addresses to override the Diddy and Dixie base palette addresses with the ones for the two currently enabled Kongs.  For the first time, no CGRAM leakage occurs with this, and Kongs properly use their palettes for various statuses.  Issues still persist. The palettes don't load properly at the start of the level yet (Diddy and Dixie's are loaded instead), but if you tag in and out, the proper palettes will load. There are also still sometimes problems with setting the palette bits for the objects.