Many map-related cosmetic changes
Code added to map animations to allow the 2 existing Kong map object to handle all four Kongs.  Donkey's map walking sprites, animation and map icon have been added.  A placeholder for Kiddy's animation and map icon have also been added. Palette loading code for maps has been modified to allow an extra palette to be loaded for Diddy and Donkey Kong.  The reds in the original Kong palette have been replaced with light blues in preparation for Kiddy's graphics.  In order to free up a line in CGRAM for the extra palette in the Crocodile Isle map, the graphic for the nose of the Crocodile Cauldron entrance has been remapped to use the smoke palette.