Fixed Kong map palette assignment issue
Code to override the Kongs' palette assignments on the map previously assumed their objects were always loaded in the $0DE2 and $0E40 slots.  However, the Crocodile Isle map is an exception.  Instead of checking if X matches $0DE2, the code now checks if it matches the pointer stored in $0597.   X is now stored to $0593 and $0597 in a different location to ensure its contents are correct by the time the check occurs.