Fix for Issue #30
Unlike the levels, only two Kong objects are loaded in the special Kong Family areas along with Klubba's Kiosk.  Objects specific to these areas load in the object entity slots $0E9E, $0EFC, $0F5A, etc.  This means the first two slots conflict with the slots used by Donkey and Kiddy in normal levels.

As a workaround, the object entities for the two enabled Kongs are loaded into object slots $0DE2 and $0E40 instead of using their typical slots in these screens.  Despite this, whichever two Kongs are loaded still load their Kong-specific variables from the same tables as in normal levels: $16B2 for Diddy, $16D8 for Dixie, $19F0 for Donkey, and $1A20 for Kiddy.  This ensures that their proper animations play.   The two disabled Kongs are not loaded at all.

The graphics for Funky's surfboard have been remapped to use the HUD numeral palette loaded in the Kong family/Klubba areas, which was already altered in a previous commit.  This needed to be done because the surfboard originally shared its palette with Dixie, who isn't always present now. Thus, the surfboard would share its palette with the Kong who was following if not for this change.

A crashing issue with Kiddy's idle animation has also been fixed.  The animation was previously using object entity variable $6C to store the return address for jumping to a sub-animations.  While this variable is valid in DKC3, it isn't in DKC2 because each object entity struct is only $5E bytes long.  $6C would correspond to variable $E (ground distance) for the next object entity, meaning the return address would frequently get overwritten when that slot is used.