Fixes for Issue #45
Thanks to some adjustments in banks 80, B4, and B8, Diddy now shows in the Cranky's Video Game Heroes screen, regardless of which Kongs are active. This is due to $0660 now being used to store the contents of kong_status instead of the contents of $08A4.  Also, in all four Kongs' main code, $0662, which is only non-zero in the Kong Family, Klubba, and Cranky's Video Game Heroes screen, is now checked and skips suspending Kongs if true.  If this check was not in place and Diddy wasn't one of the active Kongs prior to entering the Cranky's Video Game Heroes screen, Diddy would be hidden before the screen faded in.