Add No-Animal Sign softlock fix, Kiddy map sprites
Crossing a No-Animal Sign while transformed into an Animal Buddy uses the same routine used for releasing a Kong from a DK Barrel (CODE_B8A9BB).  The code loads current_sprite ($64) into X, assuming that it contains the object entity pointer of a DK Barrel in order to get the value of the Kong to release from the barrel.

However, in this case, current_sprite usually contains the pointer to a Kong object entity. As a result, it loads the Kong's object variable $4C which often happens to be zero, and stores this to the second byte of kong_status. If the first byte of kong_status is zero (Diddy), the game gets confused, as it tries to activate Diddy when he already exists, and the game engine cannot support two of the same Kong.

Kiddy's map walking animation now uses custom scaled down versions of his sprites to fit the previously created palette, instead of reusing Diddy's.

Kong palette loading code has also been changed so that palettes load in a consistent order.  Whichever Kong is leader at level load will use sprite palette line 1, and whichever one is follower will use line 2.   Previously, order was based on which Kong object was initialized first, in the order Diddy, Dixie, Donkey, and Kiddy.