Fixes for issues #54 and #58
When hitting a goal target, bits in the active Kong's object variable $30 were already manipulated to set invulnerability to enemies.  It happens that another bit in the same variable determines whether collision with Barrel Cannons is enabled.  This fix alters only two bytes in the ROM!

Also added a few lines of code to set the Kong status variable so that the gameplay demos can play properly.  As an added bonus, the Parrot Chute Panic demo has been made to play using Dixie instead of Diddy because the demo seems to have originally been recorded using her.   That demo lasts slightly longer as a result, as it doesn't end by dying to a Zinger.