Fixed Issue #36
After defeating K. Rool in either K. Rool Duel or Krocodile Kore, the game now autosaves the proper kong_status value.

In the Monkey Museum scene of both endings, things are handled a bit differently than anything else in this hack so far.  Since Cranky speaks to Diddy directly in his dialogue, he always appears in this scene as in the vanilla game. However, instead of Dixie, the second Kong who appears is the one who was being controlled at the end of the battle.  If this was already Diddy, the second Kong will be the last one who was following in the battle.