Added custom sprite graphics
Finally added custom rope animations and graphics for Donkey.
Some new files have been included in kong_hack to allow for easier implementation of custom sprite graphics and animations.
Refined image to sprite script to work correctly with different background colors. Also added a fallback for when no palette matches are found for a given pixel. The script will now find the closest color and use that instead of index 0.