Added support for Donkey and Kiddy map icons
Added a new block to save files that can store the upper bit of map icons. This was done because the original icons were stored in 2 bits. This wasn't enough to encode all the icons and the new kongs. This implementation also ensures compatibility with saves created in the original game.
Also fixed a soft lock when loading a vanilla game save.