Fix for Issues #60 and #61
Instead of checking $0662 to determine whether the player is in an NPC area (which gave false positives in certain boss levels), subroutines have been added to check various states of the NMI pointer variable ($20) and the game mode submode variable ($96).   Special thanks to p4plus2 for the suggestion of using the NMI pointer in the check.  Also, certain subroutine labels have had their names shortened.