Rewrote DK barrel cycle logic
Almost entirely rewrote the code responsible for cycling the kong number of DK barrels. This fixes issue #74 .
Added a loop that runs when freeing a kong from a DK barrel. The game now scans for other DK barrels that are loaded and forces them to cycle their kong numbers when a kong is obtained. This fixes issue #87