Added partial support for Donkey and Kiddy barrel cannons.
Updated the barrel cannon code to handle all 4 kongs.

Currently the flags for if a kong can't enter a barrel is stored in the first 2 bits of $46,x, however there are not enough free bits to also encode Donkey and Kiddy so a new sprite variable will need to be used. No suitable candidate was found with 100% confidence yet but $1E,x might been an option.

All of the known spawn scripts for barrel cannons that are Diddy/Dixie exclusive now have comments so we can more easily identify them.